Cleanup~

+++ src/main/java/vazkii/botania/common/block/dispenser/BehaviourPoolMinecart.java
@@ -10,7 +10,6 @@
  */
 package vazkii.botania.common.block.dispenser;
 
-import vazkii.botania.common.entity.EntityPoolMinecart;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockDispenser;
 import net.minecraft.block.BlockRailBase;
@@ -18,10 +17,10 @@ import net.minecraft.block.material.Material;
 import net.minecraft.dispenser.BehaviorDefaultDispenseItem;
 import net.minecraft.dispenser.IBlockSource;
 import net.minecraft.entity.item.EntityMinecart;
-import net.minecraft.item.ItemMinecart;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.world.World;
+import vazkii.botania.common.entity.EntityPoolMinecart;
 
 public class BehaviourPoolMinecart extends BehaviorDefaultDispenseItem {
 
@@ -29,9 +28,9 @@ public class BehaviourPoolMinecart extends BehaviorDefaultDispenseItem {
 	public ItemStack dispenseStack(IBlockSource p_82487_1_, ItemStack p_82487_2_) {
 		EnumFacing enumfacing = BlockDispenser.func_149937_b(p_82487_1_.getBlockMetadata());
 		World world = p_82487_1_.getWorld();
-		double d0 = p_82487_1_.getX() + (double)((float)enumfacing.getFrontOffsetX() * 1.125F);
-		double d1 = p_82487_1_.getY() + (double)((float)enumfacing.getFrontOffsetY() * 1.125F);
-		double d2 = p_82487_1_.getZ() + (double)((float)enumfacing.getFrontOffsetZ() * 1.125F);
+		double d0 = p_82487_1_.getX() + enumfacing.getFrontOffsetX() * 1.125F;
+		double d1 = p_82487_1_.getY() + enumfacing.getFrontOffsetY() * 1.125F;
+		double d2 = p_82487_1_.getZ() + enumfacing.getFrontOffsetZ() * 1.125F;
 		int i = p_82487_1_.getXInt() + enumfacing.getFrontOffsetX();
 		int j = p_82487_1_.getYInt() + enumfacing.getFrontOffsetY();
 		int k = p_82487_1_.getZInt() + enumfacing.getFrontOffsetZ();
