+++ src/main/java/vazkii/botania/common/block/mana/BlockPump.java
@@ -34,8 +34,8 @@ import vazkii.botania.common.lib.LibBlockNames;
 
 public class BlockPump extends BlockModContainer implements ILexiconable {
 
-	private static final int[] META_ROTATIONS = new int[] { 2, 5, 3, 4 }; 
-	
+	private static final int[] META_ROTATIONS = new int[] { 2, 5, 3, 4 };
+
 	public BlockPump() {
 		super(Material.rock);
 		setHardness(2.0F);
@@ -47,8 +47,8 @@ public class BlockPump extends BlockModContainer implements ILexiconable {
 
 	@Override
 	public void onBlockPlacedBy(World p_149689_1_, int p_149689_2_, int p_149689_3_, int p_149689_4_, EntityLivingBase p_149689_5_, ItemStack p_149689_6_) {
-		int l = MathHelper.floor_double((double)(p_149689_5_.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
-		p_149689_1_.setBlockMetadataWithNotify(p_149689_2_, p_149689_3_, p_149689_4_, META_ROTATIONS[l], 2); 
+		int l = MathHelper.floor_double(p_149689_5_.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
+		p_149689_1_.setBlockMetadataWithNotify(p_149689_2_, p_149689_3_, p_149689_4_, META_ROTATIONS[l], 2);
 	}
 
 	@Override
@@ -64,9 +64,9 @@ public class BlockPump extends BlockModContainer implements ILexiconable {
 
 	@Override
 	public boolean rotateBlock(World worldObj, int x, int y, int z, ForgeDirection axis) {
-        return RotationHelper.rotateVanillaBlock(Blocks.furnace, worldObj, x, y, z, axis);
+		return RotationHelper.rotateVanillaBlock(Blocks.furnace, worldObj, x, y, z, axis);
 	}
-	
+
 	@Override
 	public void registerBlockIcons(IIconRegister par1IconRegister) {
 		// NO-OP
