+++ src/main/java/vazkii/botania/common/block/subtile/generating/SubTileNightshade.java
@@ -17,7 +17,7 @@ public class SubTileNightshade extends SubTileDaybloom {
 
 	@Override
 	public int getDelayBetweenPassiveGeneration() {
-		return (int) super.getDelayBetweenPassiveGeneration();
+		return super.getDelayBetweenPassiveGeneration();
 	}
 
 	@Override
